+++ apps/run.go
@@ -255,7 +255,7 @@ func (ctx *runContext) exitCode() (int, error) {
 		return -1, errgo.Mask(err, errgo.Any)
 	}
 
-	token, err := config.C.TokenGenerator.GetAccessToken()
+	token, err := config.ScalingoClient().GetAccessToken()
 	if err != nil {
 		return -1, errgo.Notef(err, "fail to generate auth")
 	}
@@ -303,7 +303,7 @@ func (ctx *runContext) connectToRunServer() (*http.Response, net.Conn, error) {
 	if err != nil {
 		return nil, nil, errgo.Mask(err, errgo.Any)
 	}
-	token, err := config.C.TokenGenerator.GetAccessToken()
+	token, err := config.ScalingoClient().GetAccessToken()
 
 	if err != nil {
 		return nil, nil, errgo.Notef(err, "fail to generate auth")
@@ -501,7 +501,7 @@ func (ctx *runContext) uploadFile(endpoint string, file string) error {
 		return errgo.Mask(err, errgo.Any)
 	}
 
-	token, err := config.C.TokenGenerator.GetAccessToken()
+	token, err := config.ScalingoClient().GetAccessToken()
 	if err != nil {
 		return errgo.Notef(err, "fail to generate token")
 	}