+++ utils/helper.go
@@ -7,6 +7,7 @@ import (
 	"strings"
 
 	scalingo "github.com/Scalingo/go-scalingo"
+	"github.com/Scalingo/go-scalingo/http"
 
 	"gopkg.in/errgo.v1"
 )
@@ -32,11 +33,11 @@ func AskAndStopFreeTrial(c *scalingo.Client, callback func() error) error {
 // Return true if the given error is because of a Payment Required error and the free trial is
 // exceeded.
 func IsPaymentRequiredAndFreeTrialExceededError(err error) bool {
-	reqestFailedError, ok := errgo.Cause(err).(*scalingo.RequestFailedError)
+	reqestFailedError, ok := errgo.Cause(err).(*http.RequestFailedError)
 	if !ok || reqestFailedError.Code != 402 {
 		return false
 	}
-	paymentRequiredErr, ok := reqestFailedError.APIError.(scalingo.PaymentRequiredError)
+	paymentRequiredErr, ok := reqestFailedError.APIError.(http.PaymentRequiredError)
 	if !ok || paymentRequiredErr.Name != "free-trial-exceeded" {
 		return false
 	}